requirements.txt
azure-common
azure-keyvault
msrest
msrestazure
